feat: [US-003] - Make behavior-oriented API, config, and replay tests self-contained
